scripts/requirements.txt
beautifulsoup4==4.12.3
functions-framework==3.8.2
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
google-cloud-resource-manager==1.14.2
numpy==1.26.3
pandas==2.2.2
PyYAML==6.0.2
requests==2.32.3
duckdb==1.1.0
pyarrow==17.0.0
